Se corrigue anomalias multas y cobro FALTA FACTOR DEMANDA Y ESTADO TENDENCIA

+++ src/parkingsystem/Entidad/Tarifa.java
@@ -22,7 +22,7 @@ public class Tarifa {
         parking = p;
     }
 
-    public double getPrecio() {
+    public float getPrecio() {
         return precio;
     }
